+++ tests/test_controllers.py
@@ -97,16 +97,11 @@ def test_flight_controller(project: AirborneProject):
     fc = prj_ctrl.add_child(flight)
     assert hash(fc)
     assert str(fc) == str(flight)
-    assert not fc.is_active
-    prj_ctrl.activate_child(fc.uid)
-    assert fc.is_active
     assert flight.uid == fc.uid
     assert flight.name == fc.data(Qt.DisplayRole)
 
     dsc = fc.get_child(dataset.uid)
-    fc.activate_child(dsc.uid)
     assert isinstance(dsc, DataSetController)
-    assert dsc == fc.active_child
 
     dataset2 = DataSet()
     dsc2 = fc.add_child(dataset2)
@@ -114,9 +109,6 @@ def test_flight_controller(project: AirborneProject):
 
     with pytest.raises(TypeError):
         fc.add_child({1: "invalid child"})
-
-    fc.activate_child(dsc.uid)
-    assert dsc == fc.active_child
 
     fc.set_parent(None)
 
@@ -129,7 +121,6 @@ def test_flight_controller(project: AirborneProject):
 
     fc.remove_child(dsc.uid, confirm=False)
     assert 0 == len(fc.datamodel.datasets)
-    assert fc.active_child is None
 
 
 def test_FlightController_bindings(project: AirborneProject):
@@ -143,13 +134,8 @@ def test_FlightController_bindings(project: AirborneProject):
         assert 2 == len(binding)
         assert hasattr(QMenu, binding[0])
 
-    assert prj_ctrl.active_child is None
-    fc0._activate_self()
-    assert fc0 == prj_ctrl.active_child
-    assert fc0.is_active
-
     assert fc0 == prj_ctrl.get_child(fc0.uid)
-    fc0._delete_self(confirm=False)
+    fc0._action_delete_self(confirm=False)
     assert prj_ctrl.get_child(fc0.uid) is None
 
 
@@ -190,12 +176,6 @@ def test_airborne_project_controller(project):
     assert isinstance(project_ctrl.meter_model, QStandardItemModel)
     assert isinstance(project_ctrl.flight_model, QStandardItemModel)
 
-    assert project_ctrl.active_child is None
-    project_ctrl.activate_child(fc.uid)
-    assert fc == project_ctrl.active_child
-    # with pytest.raises(ValueError):
-    #     project_ctrl.activate_child(mc)
-
     project_ctrl.add_child(flight2)
 
     fc2 = project_ctrl.get_child(flight2.uid)
@@ -216,178 +196,6 @@ def test_airborne_project_controller(project):
 
     jsons = project_ctrl.save(to_file=False)
     assert isinstance(jsons, str)
-
-
-def test_dataset_controller(tmpdir):
-    """Test DataSet controls
-    Load data from HDF5 Store
-    Behavior when incomplete (no grav or traj)
-    """
-    hdf = Path(tmpdir).joinpath('test.hdf5')
-    prj = AirborneProject(name="TestPrj", path=Path(tmpdir))
-    flt = Flight("TestFlt")
-    grav_file = DataFile(DataType.GRAVITY, datetime.now(), Path(tmpdir).joinpath('gravity.dat'))
-    traj_file = DataFile(DataType.TRAJECTORY, datetime.now(), Path(tmpdir).joinpath('trajectory.txt'))
-    ds = DataSet(grav_file, traj_file)
-    seg0 = DataSegment(OID(), Timestamp.now(), Timestamp.now() + Timedelta(minutes=30), 0)
-    ds.segments.append(seg0)
-
-    flt.datasets.append(ds)
-    prj.add_child(flt)
-
-    prj_ctrl = AirborneProjectController(prj)
-    fc0 = prj_ctrl.get_child(flt.uid)
-    dsc: DataSetController = fc0.get_child(ds.uid)
-    assert 1 == dsc._segments.rowCount()
-
-    assert isinstance(dsc, DataSetController)
-    assert fc0 == dsc.get_parent()
-    assert grav_file == dsc.get_datafile(grav_file.group).datamodel
-    assert traj_file == dsc.get_datafile(traj_file.group).datamodel
-
-    grav1_file = DataFile(DataType.GRAVITY, datetime.now(), Path(tmpdir).joinpath('gravity2.dat'))
-    dsc.add_datafile(grav1_file)
-    assert grav1_file == dsc.get_datafile(grav1_file.group).datamodel
-
-    traj1_file = DataFile(DataType.TRAJECTORY, datetime.now(), Path(tmpdir).joinpath('traj2.txt'))
-    dsc.add_datafile(traj1_file)
-    assert traj1_file == dsc.get_datafile(traj1_file.group).datamodel
-
-    invl_file = DataFile('marine', datetime.now(), Path(tmpdir))
-    with pytest.raises(TypeError):
-        dsc.add_datafile(invl_file)
-
-    with pytest.raises(KeyError):
-        dsc.get_datafile('marine')
-
-    # Test Data Segment Features
-    _seg_oid = OID(tag="seg1")
-    _seg1_start = Timestamp.now()
-    _seg1_stop = Timestamp.now() + Timedelta(hours=1)
-    seg1_ctrl = dsc.add_segment(_seg_oid, _seg1_start, _seg1_stop, label="seg1")
-    seg1: DataSegment = seg1_ctrl.datamodel
-    assert _seg1_start == seg1.start
-    assert _seg1_stop == seg1.stop
-    assert "seg1" == seg1.label
-
-    assert seg1_ctrl == dsc.get_segment(_seg_oid)
-    assert isinstance(seg1_ctrl, DataSegmentController)
-    assert "seg1" == seg1_ctrl.get_attr('label')
-    assert _seg_oid == seg1_ctrl.uid
-
-    assert 2 == len(ds.segments)
-    assert ds.segments[1] == seg1_ctrl.datamodel
-    assert ds.segments[1] == seg1_ctrl.data(Qt.UserRole)
-
-    # Segment updates
-    _new_start = Timestamp.now() + Timedelta(hours=2)
-    _new_stop = Timestamp.now() + Timedelta(hours=3)
-    dsc.update_segment(seg1.uid, _new_start, _new_stop)
-    assert _new_start == seg1.start
-    assert _new_stop == seg1.stop
-    assert "seg1" == seg1.label
-
-    dsc.update_segment(seg1.uid, label="seg1label")
-    assert "seg1label" == seg1.label
-
-    invalid_uid = OID()
-    assert dsc.get_segment(invalid_uid) is None
-    with pytest.raises(KeyError):
-        dsc.remove_segment(invalid_uid)
-    with pytest.raises(KeyError):
-        dsc.update_segment(invalid_uid, label="RaiseError")
-
-    assert 2 == len(ds.segments)
-    dsc.remove_segment(seg1.uid)
-    assert 1 == len(ds.segments)
-    assert 1 == dsc._segments.rowCount()
-
-
-def test_dataset_datafiles(project: AirborneProject):
-    prj_ctrl = AirborneProjectController(project)
-    flt_ctrl = prj_ctrl.get_child(project.flights[0].uid)
-    ds_ctrl = flt_ctrl.get_child(flt_ctrl.datamodel.datasets[0].uid)
-
-    grav_file = ds_ctrl.datamodel.gravity
-    grav_file_ctrl = ds_ctrl.get_datafile(DataType.GRAVITY)
-    gps_file = ds_ctrl.datamodel.trajectory
-    gps_file_ctrl = ds_ctrl.get_datafile(DataType.TRAJECTORY)
-
-    assert grav_file.uid == grav_file_ctrl.uid
-    assert ds_ctrl == grav_file_ctrl.dataset
-    assert grav_file.group == grav_file_ctrl.group
-
-    assert gps_file.uid == gps_file_ctrl.uid
-    assert ds_ctrl == gps_file_ctrl.dataset
-    assert gps_file.group == gps_file_ctrl.group
-
-
-def test_dataset_reparenting(project: AirborneProject):
-    # Test reassignment of DataSet to another Flight
-    # Note: FlightController automatically adds empty DataSet if Flight has None
-    prj_ctrl = AirborneProjectController(project)
-    flt1ctrl = prj_ctrl.get_child(project.flights[0].uid)
-    flt2ctrl = prj_ctrl.get_child(project.flights[1].uid)
-    dsctrl = flt1ctrl.get_child(flt1ctrl.datamodel.datasets[0].uid)
-    assert isinstance(dsctrl, DataSetController)
-
-    assert 1 == len(flt1ctrl.datamodel.datasets)
-    assert 1 == flt1ctrl.rowCount()
-
-    assert 1 == len(flt2ctrl.datamodel.datasets)
-    assert 1 == flt2ctrl.rowCount()
-
-    assert flt1ctrl == dsctrl.get_parent()
-
-    dsctrl.set_parent(flt2ctrl)
-    assert 2 == flt2ctrl.rowCount()
-    assert 0 == flt1ctrl.rowCount()
-    assert flt2ctrl == dsctrl.get_parent()
-
-    # DataSetController is recreated when added to new flight.
-    assert not dsctrl == flt2ctrl.get_child(dsctrl.uid)
-    assert flt1ctrl.get_child(dsctrl.uid) is None
-
-
-def test_dataset_data_api(project: AirborneProject, hdf5file, gravdata, gpsdata):
-    prj_ctrl = AirborneProjectController(project)
-    flt_ctrl = prj_ctrl.get_child(project.flights[0].uid)
-
-    gravfile = DataFile(DataType.GRAVITY, datetime.now(),
-                        Path('tests/sample_gravity.csv'))
-    gpsfile = DataFile(DataType.TRAJECTORY, datetime.now(),
-                       Path('tests/sample_trajectory.txt'), column_format='hms')
-
-    dataset = DataSet(gravfile, gpsfile)
-
-    HDF5Manager.save_data(gravdata, gravfile, hdf5file)
-    HDF5Manager.save_data(gpsdata, gpsfile, hdf5file)
-
-    dataset_ctrl = DataSetController(dataset, flt_ctrl)
-
-    gravity_frame = HDF5Manager.load_data(gravfile, hdf5file)
-    assert gravity_frame.equals(dataset_ctrl.gravity)
-
-    trajectory_frame = HDF5Manager.load_data(gpsfile, hdf5file)
-    assert trajectory_frame.equals(dataset_ctrl.trajectory)
-
-    assert dataset_ctrl.dataframe() is not None
-    expected: DataFrame = pd.concat([gravdata, gpsdata], axis=1, sort=True)
-    expected_cols = [col for col in expected]
-
-    assert expected.equals(dataset_ctrl.dataframe())
-    assert set(expected_cols) == set(dataset_ctrl.columns)
-
-    series_model = dataset_ctrl.series_model
-    assert isinstance(series_model, QStandardItemModel)
-    assert len(expected_cols) == series_model.rowCount()
-
-    for i in range(series_model.rowCount()):
-        item: QStandardItem = series_model.item(i, 0)
-        col = item.data(Qt.DisplayRole)
-        series = item.data(Qt.UserRole)
-
-        assert expected[col].equals(series)
 
 
 def test_parent_child_activations(project: AirborneProject):
@@ -420,20 +228,3 @@ def test_parent_child_activations(project: AirborneProject):
 
     prj_ctrl.set_active(True)
     assert StateColor.ACTIVE.value == prj_ctrl.background().color().name()
-    flt_ctrl.set_active(True)
-
-    # Test exclusive/non-exclusive child activation
-    assert flt_ctrl is prj_ctrl.active_child
-    prj_ctrl.activate_child(flt2_ctrl.uid, exclusive=False)
-    assert flt_ctrl.is_active
-    assert flt2_ctrl.is_active
-
-    prj_ctrl.activate_child(flt2_ctrl.uid, exclusive=True)
-    assert flt2_ctrl.is_active
-    assert not flt_ctrl.is_active
-
-
-
-
-
-
